refactor: extract OPUS support into a mixin
The three methods under the "OPUS support methods" banner move into
src/pdsfile/_opus.py as _OpusMixin, which the class statement takes as a base in
its alphabetical slot. from_filespec and from_opus_id are byte-for-byte what they
were, and so is opus_products apart from the one edit below.

from_filespec is an alternative constructor and sits with the OPUS methods rather
than with the other from_* constructors. That is where the parent plan's PR-19
window and the file's own banner both put it, and this move keeps banner
boundaries as the boundaries.

opus_products reads the class object PdsFile, to enumerate its direct subclasses.
The Phase-5 preamble pins the pattern: the import is function-local, in the method
body, because pdsfile.py imports this module to build the class and a module-level
import would be a cycle. Those four inserted lines are the whole difference
between the moved body and the original, and they are the minimal edit that keeps
the tree importable and ruff green -- without them ruff reports F821 and the
ratchet would have to grow, which is forbidden.

defaultdict has no other user in pdsfile.py once the block leaves, but it is a
frozen member of pdsfile.pdsfile's public surface, so its import stays in the
redundant-alias re-export form. _needs_glob and abspath_for_logical_path keep
live callers there and keep their plain form.

The new module's per-file-ignores entry is a split of pdsfile.py's: RUF005 and
UP024, both already forgiven for the same lines in the same file, and every code
in pdsfile.py's entry conserves across the three files.
